task2/requirements.txt
seleniumbase
openai
langchain
python-dotenv
sentence-transformers~=2.2.1
chromadb
hns
openpyxl
pandas